requirements.txt
kivymd
kivy
python-dotenv
mysql-connector-python